moving map files into a folder
