+++ src/components/Map/MapInset.js
@@ -2,7 +2,7 @@ import React from 'react';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import Point from './Point';
-import LAYER_NAME from './MapView';
+import LAYER_NAME from '.';
 
 const mapboxgl = window.mapboxgl;
