+++ src/components/Map/Point.js
