+++ src/components/Map/style.scss
@@ -49,7 +49,7 @@
   width: 20px;
   height: 20px;
   margin: 5px;
-  background-image: url('../assets/usa.svg');
+  background-image: url('../../assets/usa.svg');
   background-size: 20px 20px;
   background-repeat: no-repeat;
   display: inline-block;
